Create renderBuilds function
